Changed templates/sundials-cvode so that all the files needed to run the integration end in .template and reaclib.py now copies the processed template files into the working directory. This leaves some source files *.f90 in templates/sundials-cvode that aren't copied to the working directory. They are useful for debugging - if you want to use them you can copy them to the working directory and (e.g. to make 'call_net' from call_net.f90) execute in the working directory:
$ make realclean
$ make program=call_net

NOTE: since reaclib.py doesn't process them, it may take some tweaking to get them to compile and run with your particular choice of reaclib rate files. But they're still useful for debugging, so I'd rather leave them in for now.

The make system uses automatic dependency resolution now. The makefiles and dep.py have been customized to permit building the debugging programs as described above and to handle the case when the working directory contains multiple Fortran program files. In such a case, make only compiles the 'program' defined in the GNUmakefile or supplied as above with 'program='.