Add nbval-ignore-output or nbval-skip tags to cells in the example notebooks that either are sensitive to roundoff error or generate plots.
For the test where we get the temperature exponent, that is sensitive
to roundoff and will likely fail, so ignore it. We already test this
feature in the pure python unit testing where we check for approximate
equality, but there isn't anything like that in the notebook test
suite nbval as far as I know.