Clarify variable names, add ydot cutoff argument
This changes variable names to clarify exactly what we're doing.

Before, `rate_value` was used to distinguish rates that weren't shown on the plot. But `rate_value` wasn't passed in as an argument anywhere.

This defines `ydot_cutoff_value` instead as an argument to `RateCollection::plot()`. The name is more accurate because we are not applying a cutoff to the rate itself, but to the ydot contribution from the rate.

Also we are not actually deleting the rates that don't satisfy this criterion from the `RateCollection` - we just are not plotting them on the graph. So this renames `deleted_rate` to `invisible_rates` to clarify that.