add support for electron captures (#69)
* add tabular_dat files

* modify rate.py for tabular rates

* modify rate.py for tabulated rates

* modiry rate.py

* modify rate-collection.py

* modify rate_collection.py

* modify rate.py

* add tabular_dat files

* modify rate.py for tabular rates

* modify rate.py for tabulated rates

* modiry rate.py

* modify rate-collection.py

* modify rate_collection.py

* modify rate.py

* add document for tabulate rate

* modify rate.py

* modify library/tabular

* modify rate_collection.py

* modify rate.py

* modify tabular data files

* xinlong

* xinlong

* updates tabular rates

* add call network_properties_init() to actual_network.F90

* Evaluate tabulated rates with (T, rho*Ye)

The `RateCollection::evaluate_rates` function was previously calling `r.eval(T, rho)`

`Rate::eval` takes `T` and `rho*Ye` for tabulated rates though, so we need to pass `rho * Ye` as the second argument.

* Clarify variable names, add ydot cutoff argument

This changes variable names to clarify exactly what we're doing.

Before, `rate_value` was used to distinguish rates that weren't shown on the plot. But `rate_value` wasn't passed in as an argument anywhere.

This defines `ydot_cutoff_value` instead as an argument to `RateCollection::plot()`. The name is more accurate because we are not applying a cutoff to the rate itself, but to the ydot contribution from the rate.

Also we are not actually deleting the rates that don't satisfy this criterion from the `RateCollection` - we just are not plotting them on the graph. So this renames `deleted_rate` to `invisible_rates` to clarify that.

* clarify comments in use_tabulated_rate notebook

* added scripts for checking the table formatting for all *-toki metadata files.

Co-authored-by: Xinlong <42697364+sailoridy@users.noreply.github.com>
Co-authored-by: Donald E. Willcox <dwillcox@users.noreply.github.com>
Co-authored-by: Donald E. Willcox <eugene.willcox@gmail.com>