Merge rp-process branch into development (#87)
* rp-process example.

* Added Jupyter notebook for rp-process network

* Added library file

* Allow RateFilter to employ user-defined functions

A callable mapping Rate -> bool may now be supplied to the RateFilter
constructor to enforce additional constraints. The parameter name is
filter_function.

* Notebook now builds an rp-process network

The network is presently rather complicated, but can be trimmed. It
excludes the pp-chain, but includes most (a, p), (p, y), and (a, y)
reactions linking nuclei from neon to nickel.

* I forgot some triple-alpha and CNO cycle rates

* Added pp-chain nuclei, prevent bintable errors

Added d, li7, be7, be8, and b8 to the core nuclei list to include
pp-chain reactions. Also added a filter to ensure that all nuclei in the
network have known binding energies.

* Add python 3.7 test ... to test.

* Switched from nucastro to python3 kernel

* Applies sorted() to sets of nuclei when building the network to make nuclei ordering deterministic.

* Added gridplot

* Modified network builder for rp-process

It now includes (a, p), (p, a), (y, p) and (y, a) reactions for heavier
nuclei.

* Some changes to plotting functions

Shifted ticks in gridplot to be cell centered, and added filter function
and rate threshold to network plot. Also set network plot axes to use
integer ticks.

* Commented out write_network call to prevent Travis build from stalling

* Added script for autogenerating rp-process networks given an endpoint

* Switched Nucleus hash function to use (Z, A) tuple

This solves the issue of Nucleus('p') not being a valid substitute for
Nucleus('h1') as a dict key. Also modified Nucleus.__eq__ to allow
(Z, A) tuples to be used as dict keys in place of Nucleus objects (e.g.
dict[1, 1] instead of dict[Nucleus('p')]).

* Added option for resolving name conflicts in rate collection

Currently, if a rate collection contains multiple rates with the same
name, the write_network call will crash with an error. This gives the
user the option of specifying a label precedence so that these conflicts
can automatically be resolved upon creating the RateCollection object.

* Added description of precedence parameter to docstring

* Removed reaclib library file

* Fix some minor numba-related things

- The numba.jitclass decorator has been moved to numba.experimental, so
  I adjusted the import to eliminate the DeprecationWarning
- Added numba to the install_requires list in setup.py

* Added nbval-skip tags to cells in rp-process notebook

Hopefully the Travis build will skip them now

* Return NotImplemented to make == commute for nuclei objects

* Incorporate Don's suggestions

* Switch back to Python 3 kernel

* Some changes to rp_process.py script

* Updated pynucastro-examples notebook with saved widget state

Co-authored-by: Donald E. Willcox <dewillcox@lbl.gov>
Co-authored-by: Donald E. Willcox <eugene.willcox@gmail.com>
Co-authored-by: Don E. Willcox <dwillcox@users.noreply.github.com>