use setuptools_scm to handle versioning (#158)
this ensures that all the versions throughout the code and doc are in sync. It pulls the version from the git tag.

We are doing a "post-release" versioning, so we will tag and then push and it will do the release version.

Also note: we will need to install before doing the sphinx build to get the version correct in the docs.