More plot options (#160)
* add reference output

* fix location

* use a relative path

* fix paths

* remove use_cse
we were using sympy's common subexpression elimination to precompute
some common factors.  But as we switch to C++, we are going to rely
on templating to assemble the network stuff, so we won't use this feature.
Also, it seems like this is something best left to the compiler in
any case.

* fix testing

* move the sympy stuff into its own module (#145)

this can help us reuse this code for the C++ nets

* move the screening factors function into the base RateCollection (#146)

since the other networks are derived from this, this allows them
to also more easily incorporate screening

* add a "secret code" to the reaclib_rates_metadata.dat (#148)

this way we can check at runtime that the correct file is being read

closes #147

* start of the C++ StarKiller nets (#149)

this covers strong / reaclib rates, but not the electron capture tables...

* use tfactors to accelerate C++ rate evaluation (#151)

* convert electron-capture tables support to C++ (#152)

* remove some unused functions in base_cxx_network.py (#156)

also fix the format of the inputs file for burn_cell

* use Legacy constants for now

* testing

* more work

* update sphinx

* this gets the version for sphinx directly from python

* install pynucastro before building the docs

* safer import

* update .gitignore -- _version.py is autogenerated

* use setuptools_scm to handle versioning (#158)

this ensures that all the versions throughout the code and doc are in sync. It pulls the version from the git tag.

We are doing a "post-release" versioning, so we will tag and then push and it will do the release version.

Also note: we will need to install before doing the sphinx build to get the version correct in the docs.

* get the full tag history on checkout

* add another example: breakout from hotCNO

* update the docs

* use amrex::Print()

* extend the plotting options

* don't return fig