Nse patch (#377)
fixed the fact that during nse calculation, evaluating partition function is invalid when the nucleus is not recorded in the table by setting the partition function to 1.0.

Added a check on the electron fraction, ye, input for nse since sometimes neutron and proton are not included in the network, which narrows down the range of valid electron fractions. The solver also diverges if the ye input is not within the valid range.

Changed the network from a uncommon complex network in the example nse notebook to subch2_full network for simplicity.