Adding new screening routine (#385)
I have included Potekhin (1998) screening routine compatible with Zhi's NSE chemical potential. Two things remain to be performed:

To implement Chugunov (2009) free energy in order to compute the NSE chemical potential and test differences.

To implement a perturbative scheme in the NSE routine to prevent overflowing. (A situation I am experimenting)

In this implementation, I plan to compute the dscor_dT and add it to this PR before it's merged.