Add gamma heating to tabular weak rates (#502)
Under certain conditions, it is favorable for the product of a weak reaction to be in an excited state (e.g. electron captures at high density). In this case, a gamma ray is emitted providing some heating. The Suzuki et al. rates do account for this and there is a column labeled "gamma-heating". From what I've looked at, this heating is generally small compared to the neutrino loss.

this PR simply adds the heating into the amrex networks, similarly to how neutrino losses are implemented.