Miscellaneous updates to pynucastro (#523)
This is a collection of updates to pynucastro that we made while implementing the reduction algorithms. See the individual commits for summaries of the separate changes. The two main updates are the addition of linking_nuclei to the RateCollection class (which we use in the algorithm implementation), and the ability to write custom libraries out to disk (which is extremely useful for testing).

Co-authored-by: Luis Rangel DaCosta <luisrd@berkeley.edu>
Co-authored-by: Eric T. Johnson <eric.t.johnson@stonybrook.edu>