Utilities used in the reduction algorithm implementations and for testing the code (#525)
This pull request includes reduction_utils.py, which is imported and used in the reduction algorithm modules.

It also includes a pair of modules that were used for testing and will be used in the example script that will be included in a later pull request. These two modules are just used locally and were not made part of the public interface. They were tested using the library generated by running ./rp_process.py te108 --write_lib rp-process-lib in the examples/rp-process directory. Running that command will require merging pull requests #523 and #524.

Co-authored-by: Luis Rangel DaCosta <luisrd@berkeley.edu>
Co-authored-by: Eric T. Johnson <eric.t.johnson@stonybrook.edu>