Loosen tabular test tolerances and don't fail on new rates (#559)
The tabular data used to generate the reference values only have 7
decimal digits of precision for the rates, and the original data from
Suzuki et. al. only have 6 digits at most, so an rtol of 1e-9 is too
strict.

Also print warnings if any rates are present in the library but not in
the test file, rather than making the test fail.