Example script with README (#529)
The script is reduction/reduction.py. It calculates the nuclear energy generation itself, so we may want to update it to use the methods provided in pynucastro. I merged most of the other pull requests into this branch since that is needed for running the script, but I can make a request with just the script and README if that is preferable.

Co-authored-by: Luis Rangel DaCosta <luisrd@berkeley.edu>