Move numpy AVX-512 workaround to conftest.py
This disables the AVX-512 routines when running pytest (both in CI and
locally), which I needed in order to reproduce the same results as
Github Actions on my machine.

It needs to go into the root of the repository because the environment
variable has to be set before numpy is first imported. If we put it in
`pynucastro/conftest.py`, pytest would import `pynucastro/__init__.py`
first.