Test the docs with pytest and nbval (#698)
note: For some reason, running inside JupyterLab produces LF ("\n") line
endings in the %cat output, while running from the terminal or in nbval
produces CRLF ("\r\n") line endings. I'm not sure why this is, but my
guess is the JupyterLab frontend is doing some sort of normalization
that the others don't.