Add bibtex citations for all references (#707)
Also fix some minor formatting issues.

See https://nbsphinx.readthedocs.io/en/latest/markdown-cells.html#Citations for the syntax for the in-notebook citations.
As a side note, the inner text of the tags gets entirely replaced with a formatted citation when building the docs. I left the original text in place, as it gets rendered weirdly in the Jupyter interface without it.