Add numpy 2.0 compatibility (#748)
As of numpy 2.0, the repr() of scalars includes the type information, which shows up in various places in interactive usage and Jupyter notebooks. Casting the result of SingleSet evaluation to a standard Python float avoids this entirely.

This also switches a reduction code test to np.sum() for compatibility with np.float64.