Add the DeBoer et al. 2017 C12(a,g)O16 rate (#769)
This creates a new alternate_rates module that can be used for adding rates that are given in papers. In this case,
the DeBoer et al. 2017 C12(a,g)O16 rate can be used as:

from pynucastro.rates import alternate_rates
c12ag_new = alternate_rates.DeBoerC12agO16()

this can be used as an alternate to the ReacLib version.

A unit test and docs page demonstrating this are also added.