change the default limits for TabularRate plots (#779)
now they are None, which means that it will use the table
limits for T and rhoY in making the plot.  The current defaults
seem to be set for the Suzuki table and do very poorly for
the Langanke rates

This also fixes how the axis tick labels are done.