fix CODATA inconsistency with Nubase and SciPy (#877)
SciPy 1.15.0 updated to CODATA 2022 constants, with both m_u and m_n
changing.  But Nubase 2020 is based on CODATA 2018.  Now we add
the old values of these constants to the pynucastro.constants module
to allow them to be used when converting mass excess into binding
energy or masses.  This ensures that the binding energy of the neutron
is 0, as it should be.  A new test was added to check this.

This also renames the spin script for consistency and updates the README

Co-authored-by: Alexander Smith <72695548+aisclark91@users.noreply.github.com>