automate the zenodo bibtex (#880)
in the docs, we now use curl to get the latest zenodo release bibtex
and literalinclude that into the sphinx source.  This way we are always
pointing to the correct version.  Also remove the explicit bibtex in the
CITATION.md and just provide the link, so we are never out of date.

Co-authored-by: Eric T. Johnson <yut23@users.noreply.github.com>