add support for specifying the order of tabular sources (#922)
When we get a TabularLibrary(), we should consider all the sources, but the same rate can be provided by
multiple sources. This adds an "ordering" keyword argument to TabularLibrary(), so we can specify the
order of precedence, e.g.,

tl = TabularLibrary(ordering=["ffn", "suzuki"])

will first read in the ffn rates and then the suzuki, replacing any ffn rates with the suzuki equivalents if both libraries provide the same decay / capture.

This also adds unit tests for the FFN rates and updates the docs to discuss the different sources

---------

Co-authored-by: Alexander Smith <soulalex@gmail.com>