add a test that directly compares ydot from C++ and python (#964)
this creates a SimpleCxxNetwork, and then make-s it and runs
it and parses the output to get the C++ dY/dt.  These are
then compared to the version from python both using the
RateCollection evaluate_ydots() and by writing the PythonNetwork
to disk as a module, importing it, and then using the rhs()
function.