when making the network plot, sort the edges before drawing (#1045)
now the edges with the highest reaction rate are plotted last,
ensuring that no other edges cross them.  This makes it easier
to see what's happening.

Also change the edges below ydot to dotted and the approx edges
to dashed (this swaps their linestyles), since this looks better.

note: we reverse sort if we are not curved