Change behavior of spins: load in spins determined by both strong and weak experimental arguments (#1115)
This changes the default behavior of the spin table. Previously, we used spins that were denoted by * or no special chars in NUBASE 2020 (doi:10.1088/1674-1137/abddae), and left off spins that were denoted by # or in parentheses. This means we included experimental information based on strong arguments and excluded information based on both theory and weak experimental arguments. In discussion, it seems that we prefer to have all spin data available, so this PR makes that the default behavior.

This will be followed up by another PR that allows us to enable/disable unreliable spins for the two interfaces that need them (NSE and deriving rates)