Allow a RateCollection or a Nucleus to restrict spins to only experimentally reliable ones (#1118)
Changed spin_table to store all spins internally, and to also store a table noting reliability of each spin. When requesting a spin, Nucleus specifies if it must be reliable or not, and spin_table either returns a spin or throws a NotImplementedError (that's the same as before). Nucleus can be chosen to be reliable through an attribute, and NSENetwork and DerivedRate can be chosen to be reliable through an initial parameter.

Also added bits to the docs and the tests.